Problemas do o free

+++ include/cgMethod.h
@@ -24,4 +24,4 @@ double normCalculator(double *r, unsigned int N);
 //	tolerance = o erro aproximado absoluto máximo (se houver)
 // 	outFileName = nome do arquivo de saida
 // =============================================================================
-int conjugateGradients(double **A, double *b, double *x, unsigned int N, unsigned int k, unsigned int maxIter, unsigned int tolerance, const char *outFileName);
+int conjugateGradients(double **A, double *b, double *x, unsigned int N, unsigned int k, unsigned int maxIter, double tolerance, const char *outFileName);
